chore: 回答例にしたがってsortメソッドをメインクラスに作った
